requirements.txt
# Core Web Framework
streamlit==1.29.0
fastapi==0.109.0
uvicorn==0.27.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Scientific Data
biopython==1.83

# Database
sqlalchemy==2.0.25

# Export Functionality
openpyxl==3.1.2
xlsxwriter==3.1.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2
